+++ src/test/java/subway/station/RestAssuredTest.java
@@ -1,4 +1,4 @@
-package subway;
+package subway.station;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
